+++ src/com/graphs/bfsDfs/CycleCheckDirectedDFS.java
@@ -1,4 +1,4 @@
-package com.learngraphs;
+package com.graphs.bfsDfs;
 
 import java.util.ArrayList;
